us ts modules in unraid-api-plugin-connect
